The following are the methods used in the
1.	Create a brightness adjusted image B1 from the image I by adding a
constant factor 50.
colorBrightness(image);

The following mathematical function is used to increase the brightness
of the image by adding a constant factor  of 50.
g(x) = alpha*f(x) + beta;

	alpha =2.0;
	beta = 50;
2.	Convert both images from RGB to HSV color space.  Write your own
code to convert from RGB to HSV color space.
convertRGBtoHSV(hsvImage);
Given an image we calculate the red, green and blue values of each
pixel that vary from 0 to 255. Using the r,g,b values we move further
in calculating the chroma component which indicates the colorfulness of
the pixel which is given by
Delta = M-m
Where,
M= max(r,g,b)
m = min(r,g,b)
The hue is the proportion of the distance around the edge of the
hexagon which passes through the projected point, originally measured
on the range [0, 1) but now typically measured in degrees [0°, 360°).
Mathematically, this definition of hue is written as,

3.	Object detection.
detectObject(objectImage);
	Convert the given RGB image to HSV color space.
	Threshold the HSV image, to keep only the red pixels
	Combine the low threshold image
	Use the Hough transform to detect circles in the combined threshold
image
	Loop over all detected circles and outline them on the original image

4.	Swapping colors from red to green
colorChange(colorImg);
For every pixel where the hue is close to red (within your threshold),
you move the hue to green.

  	color.r = image[y * w + x].g; //the green component of the image
      	color.g = image[y * w + x].r; //the red component of the image
      	color.b = image[y * w + x].b; //the blue component of the image